refactor: moved fragments into their own packages
