+++ app/src/main/java/com/northcoders/pigliotech_frontend/ui/fragments/addbook/AddBookFragment.java
@@ -1,4 +1,4 @@
-package com.northcoders.pigliotech_frontend.ui.fragments;
+package com.northcoders.pigliotech_frontend.ui.fragments.addbook;
 
 import android.os.Bundle;
