+++ app/src/main/java/com/northcoders/pigliotech_frontend/ui/fragments/signup/SignUpFragment.java
@@ -1,4 +1,4 @@
-package com.northcoders.pigliotech_frontend.ui.fragments;
+package com.northcoders.pigliotech_frontend.ui.fragments.signup;
 
 import android.os.Bundle;
 
@@ -25,6 +25,7 @@ import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.auth.UserProfileChangeRequest;
 import com.northcoders.pigliotech_frontend.R;
 import com.northcoders.pigliotech_frontend.databinding.FragmentSignUpBinding;
+import com.northcoders.pigliotech_frontend.ui.fragments.profile.ProfileFragment;
 
 
 public class SignUpFragment extends Fragment {
