+++ app/src/main/res/layout/fragment_landing_page.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".ui.fragments.LandingPageFragment">
+    tools:context=".ui.fragments.landingpage.LandingPageFragment">
 
     <Button
         android:id="@+id/btn_login"
