+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
 <FrameLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.LoginFragment">
+    tools:context=".ui.fragments.login.LoginFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
