+++ app/src/main/res/layout/fragment_profile.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".ui.fragments.ProfileFragment">
+    tools:context=".ui.fragments.profile.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
